[refactor: UserStore, Login component] remove login api call logic from UserStore
