+++ src/pages/LoginPage.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import LoginSection from 'components/Login'
 import './loginPage.scss'
+
 const LoginPage = () => {
   return (
     <div className={'container'}>
