+++ src/utils/axios.js
@@ -20,7 +20,7 @@ const axiosApi = (url, method = 'GET', data, options = {}) => {
   }
   return axios(config).then((response) => {
     const { data } = response
-    return response
+    return data
   })
 }
